Extract common websocketHandlerTest
