+++ async-http-client-backend/zio/src/test/scala/sttp/client/asynchttpclient/zio/ZioWebsocketHandlerTest.scala
@@ -1,15 +1,14 @@
 package sttp.client.asynchttpclient.zio
 
 import sttp.client._
-import sttp.client.asynchttpclient.WebSocketHandler
+import sttp.client.asynchttpclient.{AHCWebsocketHandlerTest, WebSocketHandler}
 import sttp.client.impl.zio.{TaskMonadAsyncError, convertZioIoToFuture, runtime}
 import sttp.client.monad.MonadError
 import sttp.client.testing.ConvertToFuture
-import sttp.client.testing.websocket.WebsocketHandlerTest
 import sttp.client.ws.WebSocket
 import zio.Task
 
-class ZioWebsocketHandlerTest extends WebsocketHandlerTest[Task, WebSocketHandler] {
+class ZioWebsocketHandlerTest extends AHCWebsocketHandlerTest[Task] {
   override implicit val backend: SttpBackend[Task, Nothing, WebSocketHandler] =
     runtime.unsafeRun(AsyncHttpClientZioBackend())
   override implicit val convertToFuture: ConvertToFuture[Task] = convertZioIoToFuture
@@ -17,9 +16,4 @@ class ZioWebsocketHandlerTest extends WebsocketHandlerTest[Task, WebSocketHandle
 
   override def createHandler: (Option[Int]) => WebSocketHandler[WebSocket[Task]] =
     bufferCapacity => runtime.unsafeRun(ZioWebSocketHandler(bufferCapacity))
-
-  override protected def afterAll(): Unit = {
-    backend.close().toFuture
-    super.afterAll()
-  }
 }
